requirements.txt
tensorflow==2.15.0
transformers==4.30.*
pandas==2.0.*
scikit-learn==1.3.*
Pillow==10.0.*
requests
Flask==2.3.*
numpy==1.24.*
